tasks_tracker/requirements.txt
Django
psycopg2-binary
python-dotenv
celery
kombu
gunicorn
redis
pytest
pytest-django
factory_boy
coverage
mypy
django-stubs
django-cors-headers
djangorestframework
djangorestframework-simplejwt